Renaming packages to gov.ca.water.ssrt and adding license file.
