+++ build.gradle
@@ -8,6 +8,6 @@ def versionLabel(gitInfo) {
 }
 
 allprojects {
-    group = 'gov.ca.dwr'
+    group = 'gov.ca.water'
     version = versionLabel(versionDetails())
 }
