+++ buildSrc/src/main/groovy/ssrt.java-conventions.gradle
@@ -13,3 +13,14 @@ java {
     withJavadocJar()
 }
 
+// Ensure LICENSE from the root of the repo is included in the produced jars (if present)
+tasks.withType(Jar).configureEach {
+    def licenseFile = rootProject.file('LICENSE')
+    if (licenseFile.exists()) {
+        from(licenseFile) {
+            into 'META-INF'
+            rename { 'LICENSE' }
+        }
+    }
+}
+
