+++ dwr-build-ext.gradle
@@ -1,5 +1,5 @@
 includeBuild("$gradle.ext.externalLibDir") {
     dependencySubstitution {
-        substitute(module("gov.ca.dwr:system-summary-report-tool")).using(project(":system-summary-report-tool"))
+        substitute(module("gov.ca.water:system-summary-report-tool")).using(project(":system-summary-report-tool"))
     }
 }
