+++ system-summary-report-tool/src/main/java/gov/ca/water/ssrt/Main.java
@@ -1,9 +1,8 @@
-package gov.ca.dwr.callite;
+package gov.ca.water.ssrt;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
-import java.io.IOException;
 
 import javax.swing.BoxLayout;
 import javax.swing.JButton;
