+++ system-summary-report-tool/src/main/java/gov/ca/water/ssrt/validate/VariableTolerance.java
@@ -1,4 +1,4 @@
-package gov.ca.dwr.callite.validate;
+package gov.ca.water.ssrt.validate;
 
 public class VariableTolerance {
     String variableName;
